Refactor package structure and update imports
Moved classes to more specific packages (db, model, io, service, impl) for better organization and clarity. Updated all import statements to reflect new package locations. Deleted old DefaultFruitTypeFactory and added it under model.impl. No functional changes, only structural refactoring.
