+++ src/main/java/core/basesyntax/Main.java
@@ -1,15 +1,24 @@
 package core.basesyntax;
 
-import core.basesyntax.impl.BalanceTransactionHandler;
-import core.basesyntax.impl.CsvReportSerializer;
-import core.basesyntax.impl.CsvTransactionRequestDeserializer;
-import core.basesyntax.impl.DefaultFruitTypeFactory;
-import core.basesyntax.impl.MapDatabase;
-import core.basesyntax.impl.MapTransactionService;
-import core.basesyntax.impl.PurchaseTransactionHandler;
-import core.basesyntax.impl.ReporterImpl;
-import core.basesyntax.impl.ReturnTransactionHandler;
-import core.basesyntax.impl.SupplyTransactionHandler;
+import core.basesyntax.db.Database;
+import core.basesyntax.model.Report;
+import core.basesyntax.model.Reporter;
+import core.basesyntax.model.TransactionRequest;
+import core.basesyntax.model.TransactionType;
+import core.basesyntax.service.impl.BalanceTransactionHandler;
+import core.basesyntax.io.impl.CsvReportSerializer;
+import core.basesyntax.io.impl.CsvTransactionRequestDeserializer;
+import core.basesyntax.model.impl.DefaultFruitTypeFactory;
+import core.basesyntax.db.impl.MapDatabase;
+import core.basesyntax.service.impl.MapTransactionService;
+import core.basesyntax.service.impl.PurchaseTransactionHandler;
+import core.basesyntax.model.impl.ReporterImpl;
+import core.basesyntax.service.impl.ReturnTransactionHandler;
+import core.basesyntax.service.impl.SupplyTransactionHandler;
+import core.basesyntax.io.ReportSerializer;
+import core.basesyntax.io.TransactionRequestDeserializer;
+import core.basesyntax.service.TransactionHandler;
+import core.basesyntax.service.TransactionService;
 import java.nio.file.Path;
 import java.util.HashMap;
 import java.util.List;
