--- src/main/java/core/basesyntax/impl/DefaultFruitTypeFactory.java
@@ -1,6 +1,0 @@
-package core.basesyntax.impl;
-
-import core.basesyntax.FruitTypeFactory;
-
-public final class DefaultFruitTypeFactory implements FruitTypeFactory {
-}
