+++ src/main/java/core/basesyntax/io/TransactionRequestDeserializer.java
@@ -1,4 +1,6 @@
-package core.basesyntax;
+package core.basesyntax.io;
+
+import core.basesyntax.model.TransactionRequest;
 
 import java.nio.file.Path;
 import java.util.List;
