+++ src/main/java/core/basesyntax/model/FruitType.java
@@ -1,4 +1,4 @@
-package core.basesyntax;
+package core.basesyntax.model;
 
 public record FruitType(String name) {
 }
