+++ src/main/java/core/basesyntax/model/Reporter.java
@@ -1,4 +1,6 @@
-package core.basesyntax;
+package core.basesyntax.model;
+
+import core.basesyntax.db.Database;
 
 public interface Reporter {
     Report reportWith(Database database);
