+++ src/main/java/core/basesyntax/model/impl/DefaultFruitTypeFactory.java
@@ -1,0 +1,6 @@
+package core.basesyntax.model.impl;
+
+import core.basesyntax.model.FruitTypeFactory;
+
+public final class DefaultFruitTypeFactory implements FruitTypeFactory {
+}
