+++ src/main/java/core/basesyntax/service/TransactionHandler.java
@@ -1,4 +1,7 @@
-package core.basesyntax;
+package core.basesyntax.service;
+
+import core.basesyntax.model.FruitType;
+import core.basesyntax.db.Database;
 
 import java.math.BigInteger;
