+++ src/main/java/core/basesyntax/service/TransactionService.java
@@ -1,4 +1,6 @@
-package core.basesyntax;
+package core.basesyntax.service;
+
+import core.basesyntax.model.TransactionType;
 
 public interface TransactionService {
     TransactionHandler selectHandler(TransactionType transactionType);
