+++ src/main/java/core/basesyntax/service/impl/BalanceTransactionHandler.java
@@ -1,8 +1,8 @@
-package core.basesyntax.impl;
+package core.basesyntax.service.impl;
 
-import core.basesyntax.Database;
-import core.basesyntax.FruitType;
-import core.basesyntax.TransactionHandler;
+import core.basesyntax.db.Database;
+import core.basesyntax.model.FruitType;
+import core.basesyntax.service.TransactionHandler;
 import java.math.BigInteger;
 
 public class BalanceTransactionHandler implements TransactionHandler {
