+++ src/main/java/core/basesyntax/service/impl/MapTransactionService.java
@@ -1,8 +1,8 @@
-package core.basesyntax.impl;
+package core.basesyntax.service.impl;
 
-import core.basesyntax.TransactionHandler;
-import core.basesyntax.TransactionService;
-import core.basesyntax.TransactionType;
+import core.basesyntax.service.TransactionHandler;
+import core.basesyntax.service.TransactionService;
+import core.basesyntax.model.TransactionType;
 import java.util.Map;
 
 public class MapTransactionService implements TransactionService {
